chore: moved models and lists in to common folder
