+++ src/common/appConfiguration.js
