+++ src/common/packageCodeLensList.js
@@ -2,9 +2,9 @@
  *  Copyright (c) Peter Flannery. All rights reserved.
  *  Licensed under the MIT License. See LICENSE in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
-import {assertDefined} from '../common/typeAssertion';
 import {Range, Uri} from 'vscode';
-import {PackageCodeLens} from '../models/packageCodeLens';
+import {assertDefined} from './typeAssertion';
+import {PackageCodeLens} from './packageCodeLens';
 
 export class PackageCodeLensList {
